requirements.txt
asgiref==3.7.2
Django==4.2.8
django-cors-headers==4.3.1
django-extensions==3.2.3
django-crispy-forms==1.12.0
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
crispy-bootstrap5==0.1.0
djangorestframework==1.11.0